+++ library/objective-c/BUILD
@@ -1,20 +1,11 @@
 licenses(["notice"])  # Apache 2
 
-exports_files(["EnvoyEngine.h"])
-
-filegroup(
-    name = "envoy_engine_hdrs",
-    srcs = ["EnvoyEngine.h"],
-    visibility = ["//visibility:public"],
-)
+exports_files(["EnvoyEngine.h", "EnvoyTypes.h"])
 
 objc_library(
     name = "envoy_engine_objc_lib",
     srcs = [
         "EnvoyEngine.mm",
-    ],
-    hdrs = [
-        "EnvoyEngine.h",
     ],
     copts = ["-std=c++14"],
     visibility = ["//visibility:public"],
